+++ src/EmailValidation.Console/CsvFileProcessor.cs
@@ -4,6 +4,7 @@ using System.Text;
 using CsvHelper;
 using CsvHelper.Configuration;
 using EmailValidation.Core;
+using EmailValidation.Application;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.Extensions.DependencyInjection;